+++ src/endpoints/index.ts
@@ -5,16 +5,23 @@ import fs from "fs";
 import client_subsequentRequests from "./next_js/client_subsequentRequests.js";
 import client_jsFilesHref from "./next_js/client_jsFilesHref.js";
 import client_jsonParse from "./next_js/client_jsonParse.js";
-import getWebpacks from "./next_js/getWebpacks.js";
+import client_mappedJsonFile from "./next_js/client_mappedJsonFile.js";
 
 // Report Generation
-import gen_markdown from "./gen_report/gen_markdown.js";
 import gen_json from "./gen_report/gen_json.js";
 
 const techs = ["Next.JS (next)"];
-const outputFormats = ["md", "json"];
+const outputFormats = ["json"];
 
-const endpoints = async (url, directory, output, outputFormat, tech, list, subsequentRequestsDir) => {
+const endpoints = async (
+    url: string | undefined,
+    directory: string | undefined,
+    output: string | undefined,
+    outputFormat: string[] | undefined,
+    tech: string | undefined,
+    list: boolean | undefined,
+    mappedJsonFile: string | undefined
+) => {
     console.log(chalk.cyan("[i] Loading endpoints module"));
 
     // list available technologies
@@ -32,12 +39,6 @@ const endpoints = async (url, directory, output, outputFormat, tech, list, subse
             console.log(chalk.red("[!] Invalid output format"));
             return;
         }
-    }
-
-    // check if the directory is present
-    if (!directory) {
-        console.log(chalk.red("[!] Please provide a directory"));
-        return;
     }
 
     // check if the technology is present
@@ -63,36 +64,33 @@ const endpoints = async (url, directory, output, outputFormat, tech, list, subse
     if (tech === "next") {
         console.log(chalk.cyan("[i] Checking for client-side paths for Next.JS"));
 
-        // check if the subsequent requests directory is present
-        if (!subsequentRequestsDir) {
-            console.log(
-                chalk.red("[!] Please provide a directory containing subsequent requests (--subsequent-requests-dir)")
-            );
-            return;
-        }
-
-        // check if the subsequent requests directory exists
-        if (!fs.existsSync(subsequentRequestsDir)) {
-            console.log(chalk.red("[!] Directory containing subsequent requests does not exist"));
-            return;
-        }
-
+        // var to store all the paths found
         let final_client_side: string[] = [];
-        const client_subsequentRequestsResult = await client_subsequentRequests(subsequentRequestsDir, url);
-        final_client_side.push(...client_subsequentRequestsResult);
 
-        // first, get all the webpacks
-        // const webpacksFound = await getWebpacks(directory);
+        if (directory) {
+            const subsequentRequestsDir = directory + "/___subsequent_requests";
+            // check if the subsequent requests directory exists
+            if (!fs.existsSync(subsequentRequestsDir)) {
+                console.log(chalk.red("[!] Directory containing subsequent requests does not exist"));
+                return;
+            }
 
-        const client_jsFilesHrefResult = await client_jsFilesHref(directory);
-        final_client_side.push(...client_jsFilesHrefResult);
+            const client_subsequentRequestsResult = await client_subsequentRequests(subsequentRequestsDir, url);
+            final_client_side.push(...client_subsequentRequestsResult);
 
-        const client_jsonParseResult = await client_jsonParse(directory);
-        final_client_side.push(...client_jsonParseResult);
+            const client_jsFilesHrefResult = await client_jsFilesHref(directory);
+            final_client_side.push(...client_jsFilesHrefResult);
 
-        if (outputFormat.includes("md")) {
-            const gen_markdownResult = await gen_markdown(url, final_client_side, output);
+            const client_jsonParseResult = await client_jsonParse(directory);
+            final_client_side.push(...client_jsonParseResult);
         }
+
+        // now, use the mapped JSON file to find more paths
+        if (mappedJsonFile) {
+            const client_mappedJsonFileResult = await client_mappedJsonFile(mappedJsonFile);
+            final_client_side.push(...client_mappedJsonFileResult);
+        }
+
         if (outputFormat.includes("json")) {
             const gen_jsonResult = await gen_json(url, final_client_side, output);
         }